Accurate error tracker implemented, line validator function introduced to identify valid phrases on each line and identify those lines that are not empty ( fill with trash ) and the lines that contains tokens, but the phrase is not recognised by the compiler as valid and this will generate an error
The full report if you want to clearly see what's changed is the main
flow that the compiler uses to compile the source code. Before this
commit the compiler was just iterating to each token and formulate the
desired expression as soon the right combination was hitter, this is
cool but doesn't track invalid line or empty lines and failed to
identify clearly the point of failure during the compilation and the
invalid statement or phrases in the code. So instead of a token-related
iteration the compiler shifter to a line related operation with a
validation to ensure the alignment with the token tracker of the lexer.
This allow the compiler to keep track of wich line is compiling, see if
the current line is empty, has tokens or is just fill with shit, and by
identify this is possible to create an accurate tracking of the
statement of the compilation. Notice that this is still dangerous
because is simple to disalign the line tracker from the token tracker if
the expression is not parsed in a single line. Essentially functions
related to the expression parsing ( like the function expression parsing
) must just parse the line in wich the function is defined, after that
the next line switching must be done by the compiler routine, if this is
followed the tracker cannot be broken, but if the token is increased to
much to result in a different line identifier compared to the main line
identifier ( or tracker ) of the compiler routine the compilation of the
statement can infact broke and result in an invalid statement.