Better integration of the compilation context inside the compiler structure and automatic context switch implemented
The automatic context switch still maintain the line separation
structure introduced by the previows commits, the only thing that change
is that now is possible to identify when a new line occurs or not inside
the token concatenation. This allow the compiler to change the way to
identify an expression allowing a "multi-line" composition ( like a huge
if case ) from single line composition ( like function declaration and
such ) returning a "TOKEN_TABLE_END" token from the
XEB_GET_CURRENT_TOKEN() if there is no tokens left on the current
context immediately after the context switching.